Add cloud_firestore_mocks and unit test with real service Fixes #233
