+++ pubspec.yaml
@@ -23,8 +23,7 @@ dev_dependencies:
   test:
   flutter_test:
     sdk: flutter
-  mockito: ^4.1.1
-
+  mockito:
   flutter_launcher_icons:
 
 flutter:
