+++ test/utils/widget_test_utils.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+Widget wrapWidget(Widget child) {
+  return MaterialApp(
+    home: Scaffold(
+      body: child,
+    ),
+  );
+}
