+++ test/widgets/chat_page_test.dart
@@ -1,15 +1,13 @@
-import 'package:adventures_in_chat_app/widgets/messages/chat_page.dart';
 import 'package:adventures_in_chat_app/models/conversation_item.dart';
 import 'package:adventures_in_chat_app/models/message.dart';
-import 'package:adventures_in_chat_app/services/database_service.dart';
 import 'package:adventures_in_chat_app/widgets/messages/chat_message.dart';
+import 'package:adventures_in_chat_app/widgets/messages/chat_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
-class MockDatabaseService extends Mock implements DatabaseService {
-  MockDatabaseService();
-}
+import '../mocks/database_mocks.dart';
+import '../utils/widget_test_utils.dart';
 
 void main() {
   group('ChatMessage', () {
@@ -127,12 +125,4 @@ void main() {
       ));
     });
   });
-}
-
-Widget wrapWidget(Widget child) {
-  return MaterialApp(
-    home: Scaffold(
-      body: child,
-    ),
-  );
 }
